Experiments/Bots/Bot_1_and_2/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
audio-recorder-streamlit==0.0.9
awscli==1.32.18
blinker==1.8.2
botocore==1.34.18
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
deepgram-sdk==3.1.4
distro==1.9.0
docutils==0.16
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-groq==0.0.1
langchain-openai==0.0.7
langsmith==0.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.9.1
pygame==2.5.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.7.2
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==2.0.27
stack-data==0.6.3
streamlit==1.29.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.7
validators==0.28.3
verboselogs==1.7
wcwidth==0.2.13
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.2
